Maps the menu button to the analog button.
Allows the user to toggle the analog stick with the menu button.  Fix
for Sky Odyssey.

+++ pcsx2/PAD/Linux/KeyStatus.h
@@ -31,9 +31,9 @@ class KeyStatus
 private:
 	const u8 m_analog_released_val;
 
-	u16 m_button[GAMEPAD_NUMBER];
-	u16 m_internal_button_kbd[GAMEPAD_NUMBER];
-	u16 m_internal_button_joy[GAMEPAD_NUMBER];
+	u32 m_button[GAMEPAD_NUMBER];
+	u32 m_internal_button_kbd[GAMEPAD_NUMBER];
+	u32 m_internal_button_joy[GAMEPAD_NUMBER];
 
 	u8 m_button_pressure[GAMEPAD_NUMBER][MAX_KEYS];
 	u8 m_internal_button_pressure[GAMEPAD_NUMBER][MAX_KEYS];
@@ -62,7 +62,7 @@ public:
 	void press(u32 pad, u32 index, s32 value = 0xFF);
 	void release(u32 pad, u32 index);
 
-	u16 get(u32 pad);
+	u32 get(u32 pad);
 	u8 get(u32 pad, u32 index);
